+++ src/config.js
@@ -3,6 +3,7 @@
 export const Agents = {}
 export const Epics = {}
 export const NewId = []
+export const Types = {}
 export const ReducerForKey = []
 export const ViewReducer = []
 export const MetaEnhancers = [